# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agora API",
    "version": "0.1.0",
    "description": "Industrial-grade marketplace SaaS — API contracts (ADR-0011)."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/v1/auth/mfa/setup": {
      "post": {
        "operationId": "mfaSetup",
        "summary": "Start MFA enrollment (TOTP) — returns secret + otpauth URI",
        "tags": [
          "auth",
          "mfa"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Secret + otpauth URI"
          },
          "409": {
            "description": "Already enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/mfa/enable": {
      "post": {
        "operationId": "mfaEnable",
        "summary": "Confirm enrollment (password + code); issues recovery codes once",
        "tags": [
          "auth",
          "mfa"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Enabled + recovery codes"
          }
        }
      }
    },
    "/v1/auth/mfa/verify": {
      "post": {
        "operationId": "mfaVerify",
        "summary": "Complete login with TOTP code or recovery code",
        "tags": [
          "auth",
          "mfa"
        ],
        "responses": {
          "200": {
            "description": "Token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Bad code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/mfa/disable": {
      "post": {
        "operationId": "mfaDisable",
        "summary": "Disable MFA (password + code); revokes session family",
        "tags": [
          "auth",
          "mfa"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Disabled"
          }
        }
      }
    },
    "/v1/auth/mfa/recovery": {
      "post": {
        "operationId": "mfaRecovery",
        "summary": "Regenerate recovery codes (password + code)",
        "tags": [
          "auth",
          "mfa"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "New recovery codes"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code, e.g. EMAIL_TAKEN, INVALID_CREDENTIALS",
                "example": "EMAIL_TAKEN"
              },
              "message": {
                "type": "string",
                "description": "Human-readable message",
                "example": "Email is already registered"
              },
              "requestId": {
                "type": "string",
                "description": "Correlation id (x-request-id)"
              },
              "details": {
                "type": "object",
                "additionalProperties": {},
                "description": "Optional field-level details"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuthLoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "RS256 JWT, 15 min TTL"
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque refresh token, 30 d TTL, rotated on use"
          },
          "expiresIn": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "description": "Access token TTL in seconds",
            "example": 900
          },
          "tokenType": {
            "type": "string",
            "const": "Bearer"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "UUID v7 identifier",
                "example": "01961e78-4d99-7c20-8000-000000000000"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "mfaRequired": {
                "type": "boolean"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "id",
              "email",
              "mfaRequired",
              "roles"
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "tokenType",
          "user"
        ]
      },
...
    }
  }
}